+++ shared/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.58] - 2023-05-26
+
+- Added a comment on session.proto regarding use of maps for session's actors
+
 ## [0.0.0.57] - 2023-05-04
 
 - Move `utility/types/errors.go` into `shared/core/types/`